Add attention op to linalg_ext (#11928)
This patch adds the attention op to the linalg ext dialect. It also adds a lowering that decomposes the attention into a sequence of linalg generic ops (batch matmul + softmax + batch matmul). Finally, we add the pass into the CPU and CUDA pipeline along with an e2e test to validate functional accuracy.