Remove conditionals restricting workflows to a specific repo (#12012)
This was to prevent workflows, especially scheduled ones, from running
in people's forks. GitHub now has functionality for fork owners to
disable a workflow and scheduled workflows are disabled by default in
forks. Having this repeated in every job is extra cruft and there are
cases where a fork owner who's using the fork as a site of active
development (as opposed to just for upstreaming) may want to do their
own releases and such and pull from upstream without maintaining a
patch. It seems like having everyone use the repository option is the
better way to support this.


https://docs.github.com/en/actions/managing-workflow-runs/disabling-and-enabling-a-workflow

skip-ci: Doesn't affect CI workflows.