Redo handling of operand reuse to prevent stack allocation. (#12041)
For dispatches with

```
%r = <some compute op>
%store = linalg.generic ... ins(%r, ...) outs(%empty :)
```
where the value of `%store` is the final result of the dispatch, to avoid creation of stacks for `%r` when vectorization does not kick in, the user is converted to

```
%store = linalg.generic ... ins(, ...) outs(%r :)
```
This is a very specific change to work around something that really should be handled later on during bufferization (but is outside scope of bufferization today). This change needs to kick in very specific conditions (since its a work-around, and really the created linalg.generic is somewhat of an anti-pattern), using a pattern rewrite to make this change does not allow controlling the transformation. Change this to use a walk and trigger only when needed.

Fixes #12019.

Fixes https://github.com/iree-org/iree/issues/12019