Enable Python release building in CI. (#12346)
Per #12341, we should have a smoke test of python package building. Note
that this is still a proxy for the actual workflow that does this, as
that differs in setup, metadata, versions built and schedule and is
unsuitable for direct inclusion. However, the "inside" is all scripted
and invariant and run identically between the CI and release building.
This should reduce the rate of breaking incidents.

The release job also builds the "main" package tarball, which is not yet
included here and has the potential to break in novel ways (but
historically is much easier to predict/test).

This also would enable, should we ever want to, the ability to have CI
jobs *depend on* just the built Python packages, using them to do
further testing and automation without rebuilding entire parts of the
project. That is left for a future exercise.

Fixes #12341 (modulo future enhancements related to tarball build testing, etc).