[LLVMGPU] Enable tensor.pack op e2e execution on cuda. (#12370)
It takes 2.7 ms to finish the tensor.pack e2e tests. This is because the
fully dynamic tensor.pack test cases are not vectorized. There are two
fully dynamic cases in total. If we exclude the test, it takes ~0.57 ms.
The pipeline is not well tuned yet, it is mostly copied from LLVMCPU and
for correctness tests.

It also splits the pack.mlir test file into two files. One has static
inner tiling sizes, and the other has dynamic inner tiling sizes. The
dynamic cases are only used by microkernels, which aims to have unify
kernel for all the cases. Currently, the cuda side does not support it
because it introduces unbounded memory allocation.