Build libIREECompiler.so in bazel. (#12507)
* Moves the deeply nested build rule to lib/. Unlike CMake, Bazel offers
no control on where things are output in the tree (or name), so we just
put the binary target in the place where CMake puts it, preserving the
directory layout of artifacts between them. This results in a small
deviation on where this target is actually declared (lib/ in bazel and
iree/compiler/API2:SharedImpl in CMake) but the same output path.
* Wires up the loader_test in bazel and makes it able to find the shared
library data dependency.
* Extends bazel_to_cmake to respect the `skip-bazel_to_cmake` tag to
make it skip conversion of common targets. I've needed this for some
time and I expect will let us normalize more build files that only
differ by a target or two.
* I've confirmed that this works in Google's internal build system (with
appropriate buildozer path changes and such).