Add C APIs sufficient for tunneling LLVM command line handling into derived tools. (#12710)
This does some delicate surgery to rationalize the normal LLVM
initialization sequence and make it suitable for API usage for the
limited use case of subordinating IREE compiler's command line
processing to an outer component (i.e. by having "--" style separators
or "-Xcompiler" like options in another command line library).

`iree-compile` itself still uses its privileged position as built-in to
actually define new global command line options, which the C API does
not expose. This works because it is always statically linked into the
main libIREECompiler.so shared library.

BREAKING C API CHANGE: `ireeCompilerGlobalInitialize()` now takes no
arguments whereas it used to take a boolean "usesCommandLine" option
which was a half-hearted attempt to make this case work previously. The
only use of this as true was in-tree and it wasn't really possible to
use right externally. If passing false, just drop the argument. If
passing true, use a combination of ireeCompilerGetProcessCLArgs and
ireeCompilerSetupGlobalCL as shown in iree_compile_lib.cc. Also, the
sequencing has been reworked to allow nested calls to
initialize/shutdown. The lack of this was causing unreliability in the
PJRT client, which is the major out of tree user of the API. Given the
low use and broken state of this API, it is preferable to just clean it
up now vs retain compatibility.