Integrate the work for hoisting address computation (#12643)
Add a pass to extract address computation from memref.load and
nvgpu.ldmatrix.
Plumb the affine.apply decomposition through a new pass:
decompose-affine-ops.
Rework the lowering pipeline to connect all the piece together:
1. extract-address-computation turns address computation into subviews
2. expand-strided-metadata turns subviews into affine.apply
3. licm hoists the code introduced by # 2 in the right scf.for loop
4. decompose-affine-ops breaks down the `affine.apply`s so that the
resulting
   subexpressions can be hoisted in the right loops.
5. licm hoists the code introduced by # 4
6. lower-affine materializes the decomposed `affine.apply`s. We do that
early
   to avoid the canonicalization to undo this work.

Phase 3-5 needs to run on `scf.for`, so the whole process has to run
before
scf to cf.

TODO:
    - Add support for vector.transfer_xxx

Note: extract-address-computation could be moved to LLVM open source,
but we need to figure out where it could live since it has both a
dependency on
memref and nvgpu. We probably want to come up with an interface like
`isAddressComputationExtractable` to push it upstream.