Use benchmark name from benchmark config object (#12723)
For running new benchmark suite, use the benchmark name from benchmark
config instead of generating from `BenchmarkInfo` and `CompilationInfo`.
This allows us to use the new benchmark suite as the SoT of all
benchmark configurations.

Perf dashboard is using benchmark id as primary key for new benchmark
suites, so this change won't break the series. Android benchmarks still
use the legacy name as key and this change doesn't touch it.