[Transform] Add op to apply im2col patterns and update the workgroup count region (#12804)
Required for implementing a pipeline for implicit gemm on GPU backends.
Im2col converts a convolution to a gather on the input and a gemm. This
collapses the image dimensions of the output for 2D convolutions and
thus the workgroup count compute region needs to be updated to match the
new iteration ranges for the main compute op.