Add E2E MNIST training test (#12017)
Add a test that verifies the accuracy of performing a training step on
an MNIST model. The model was exported using IREE Jax. Along with the
test there are some facilities for Python testing that are included to
be used by other future tests.

This test reproduces the problem at
https://github.com/iree-org/iree/issues/11991 but without the dependency
to Jax and IREE Jax.

I am not sure that it is correct to add the model and the expected
output data to this source tree. They are 2.5 MB in total.

---------

Co-authored-by: Boian Petkantchin <boian@nod-labs.com>
Co-authored-by: Geoffrey Martin-Noble <gcmn@google.com>